OSD: Implement the NotifyingLock interface on Session
Provide the Session an OSD pointer, and implement a notify() function. Use it
to queue up a new C_SessionWaker on the op_gen_wq, which will call
dispatch_session_waiting() when it's executed. This means that we can avoid
processing requests right when they come in, while easily preserving ordering
(we just don't pop the OpRequest off of the waiting_for_map list), and
scheduling a wakeup that keeps us from accidentally sleeping forever.

Signed-off-by: Greg Farnum <greg@inktank.com>