OSD: enable lock retries on fast dispatch
Use try_get_pg_or_queue_for_pg in handle_op(), so that we cannot deadlock
(although we can still block for short periods of time) on other threads
in the OSD doing work. This is easy to flip on -- we simply return the success
value, and the entire chain up to dispatch_session_waiting behaves
appropriately. The NotifyingLock and associated interfaces we just
implemented, meanwhile, ensure that the Session will get a wakeup regardless
of anything else that happens.

Signed-off-by: Greg Farnum <greg@inktank.com>