OSD:check_replay_queue: do not drop non-active replay pgs on the floor
Fixes: #8758
Backport: firefly, dumpling
Signed-off-by: Samuel Just <sam.just@inktank.com>
Reviewed-by: Samuel Just <sam.just@inktank.com>

If a PG is in replay but it's not active yet by the time the OSD tick
gets to running the replay queue, we'll drop the PG from the replay
queue without clearing its replay state.

This patch arranges for PGs that weren't ready for replaying (or that
weren't replayed for any other reason) to be requeued, so that they
will be retried in the next tick.

I have confirmed, with this patch, that my memory-starved nodes would
often trigger the new log message, indicating the PGs were not active
yet.  I have not observed instances of non-primary or replay_until
mismatch, and I'm not sure the latter should justify a requeue, an
assertion failure or whatever, but simply dropping the PG on the floor
leaving it indefinitely in replay state is not something that should
be done, since it prevents other operations from being processed.

This patch does NOT fix the other problem reported in Issue #8758,
that restarting any of the osds holding a PG is enough to clear the
replay state.

Signed-off-by: Alexandre Oliva <oliva@gnu.org>